forex-ai-bot/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
ta==0.11.0
openai==0.28.1
python-telegram-bot==20.4
pyyaml==6.0
